replace u32 offsets with u64 and remove alignment module
- delete align.rs and remove 16-byte alignment system
- change all offset/position types from u32 to u64 throughout storage crate
- update archive benchmarks to use 64KB values (Vec<u8>)
- consolidate redundant truncation tests and fix misleading comments
- regenerate conformance hashes for new on-disk format

+++ storage/src/freezer/storage.rs
@@ -30,15 +30,15 @@ const RESIZE_THRESHOLD: u64 = 50;
 #[derive(Hash, PartialEq, Eq, PartialOrd, Ord, Clone, Copy)]
 #[cfg_attr(feature = "arbitrary", derive(arbitrary::Arbitrary))]
 #[repr(transparent)]
-pub struct Cursor([u8; u64::SIZE + u32::SIZE + u32::SIZE]);
+pub struct Cursor([u8; u64::SIZE + u64::SIZE + u64::SIZE]);
 
 impl Cursor {
     /// Create a new [Cursor].
-    fn new(section: u64, offset: u32, size: u32) -> Self {
-        let mut buf = [0u8; u64::SIZE + u32::SIZE + u32::SIZE];
+    fn new(section: u64, offset: u64, size: u64) -> Self {
+        let mut buf = [0u8; u64::SIZE + u64::SIZE + u64::SIZE];
         buf[..u64::SIZE].copy_from_slice(&section.to_be_bytes());
-        buf[u64::SIZE..u64::SIZE + u32::SIZE].copy_from_slice(&offset.to_be_bytes());
-        buf[u64::SIZE + u32::SIZE..].copy_from_slice(&size.to_be_bytes());
+        buf[u64::SIZE..u64::SIZE + u64::SIZE].copy_from_slice(&offset.to_be_bytes());
+        buf[u64::SIZE + u64::SIZE..].copy_from_slice(&size.to_be_bytes());
         Self(buf)
     }
 
@@ -48,13 +48,13 @@ impl Cursor {
     }
 
     /// Get the offset of the cursor.
-    fn offset(&self) -> u32 {
-        u32::from_be_bytes(self.0[u64::SIZE..u64::SIZE + u32::SIZE].try_into().unwrap())
+    fn offset(&self) -> u64 {
+        u64::from_be_bytes(self.0[u64::SIZE..u64::SIZE + u64::SIZE].try_into().unwrap())
     }
 
     /// Get the size of the value.
-    fn size(&self) -> u32 {
-        u32::from_be_bytes(self.0[u64::SIZE + u32::SIZE..].try_into().unwrap())
+    fn size(&self) -> u64 {
+        u64::from_be_bytes(self.0[u64::SIZE + u64::SIZE..].try_into().unwrap())
     }
 }
 
@@ -62,7 +62,7 @@ impl Read for Cursor {
     type Cfg = ();
 
     fn read_cfg(buf: &mut impl Buf, _: &Self::Cfg) -> Result<Self, commonware_codec::Error> {
-        <[u8; u64::SIZE + u32::SIZE + u32::SIZE]>::read(buf).map(Self)
+        <[u8; u64::SIZE + u64::SIZE + u64::SIZE]>::read(buf).map(Self)
     }
 }
 
@@ -73,7 +73,7 @@ impl CodecWrite for Cursor {
 }
 
 impl FixedSize for Cursor {
-    const SIZE: usize = u64::SIZE + u32::SIZE + u32::SIZE;
+    const SIZE: usize = u64::SIZE + u64::SIZE + u64::SIZE;
 }
 
 impl commonware_utils::Span for Cursor {}
@@ -193,7 +193,7 @@ struct Entry {
     // Section in which this slot was written
     section: u64,
     // Position in the key index for this section
-    position: u32,
+    position: u64,
     // Number of items added to this entry since last resize
     added: u8,
     // CRC of (epoch | section | position | added)
@@ -205,7 +205,7 @@ impl Entry {
     const FULL_SIZE: usize = Self::SIZE * 2;
 
     /// Compute a checksum for [Entry].
-    fn compute_crc(epoch: u64, section: u64, position: u32, added: u8) -> u32 {
+    fn compute_crc(epoch: u64, section: u64, position: u64, added: u8) -> u32 {
         let mut hasher = crc32fast::Hasher::new();
         hasher.update(&epoch.to_be_bytes());
         hasher.update(&section.to_be_bytes());
@@ -215,7 +215,7 @@ impl Entry {
     }
 
     /// Create a new [Entry].
-    fn new(epoch: u64, section: u64, position: u32, added: u8) -> Self {
+    fn new(epoch: u64, section: u64, position: u64, added: u8) -> Self {
         Self {
             epoch,
             section,
@@ -237,7 +237,7 @@ impl Entry {
 }
 
 impl FixedSize for Entry {
-    const SIZE: usize = u64::SIZE + u64::SIZE + u32::SIZE + u8::SIZE + u32::SIZE;
+    const SIZE: usize = u64::SIZE + u64::SIZE + u64::SIZE + u8::SIZE + u32::SIZE;
 }
 
 impl CodecWrite for Entry {
@@ -255,7 +255,7 @@ impl Read for Entry {
     fn read_cfg(buf: &mut impl Buf, _: &Self::Cfg) -> Result<Self, commonware_codec::Error> {
         let epoch = u64::read(buf)?;
         let section = u64::read(buf)?;
-        let position = u32::read(buf)?;
+        let position = u64::read(buf)?;
         let added = u8::read(buf)?;
         let crc = u32::read(buf)?;
 
@@ -271,32 +271,32 @@ impl Read for Entry {
 
 /// Sentinel value indicating no next entry in the collision chain.
 const NO_NEXT_SECTION: u64 = u64::MAX;
-const NO_NEXT_POSITION: u32 = u32::MAX;
+const NO_NEXT_POSITION: u64 = u64::MAX;
 
 /// Key entry stored in the segmented/fixed key index journal.
 ///
 /// All fields are fixed size, enabling efficient collision chain traversal
 /// without reading large values.
 ///
-/// The `next` pointer uses sentinel values (u64::MAX, u32::MAX) to indicate
+/// The `next` pointer uses sentinel values (u64::MAX, u64::MAX) to indicate
 /// "no next entry" instead of Option, ensuring fixed-size encoding.
 #[derive(Debug, Clone, PartialEq)]
 struct KeyEntry<K: Array> {
     /// The key for this entry.
     key: K,
     /// Pointer to next entry in collision chain (section, position in key index).
-    /// Uses (u64::MAX, u32::MAX) as sentinel for "no next".
+    /// Uses (u64::MAX, u64::MAX) as sentinel for "no next".
     next_section: u64,
-    next_position: u32,
-    /// Aligned offset in value journal (same section).
-    value_offset: u32,
+    next_position: u64,
+    /// Byte offset in value journal (same section).
+    value_offset: u64,
     /// Size of value data in the value journal.
-    value_size: u32,
+    value_size: u64,
 }
 
 impl<K: Array> KeyEntry<K> {
     /// Create a new [KeyEntry].
-    fn new(key: K, next: Option<(u64, u32)>, value_offset: u32, value_size: u32) -> Self {
+    fn new(key: K, next: Option<(u64, u64)>, value_offset: u64, value_size: u64) -> Self {
         let (next_section, next_position) = next.unwrap_or((NO_NEXT_SECTION, NO_NEXT_POSITION));
         Self {
             key,
@@ -308,7 +308,7 @@ impl<K: Array> KeyEntry<K> {
     }
 
     /// Get the next entry in the collision chain, if any.
-    const fn next(&self) -> Option<(u64, u32)> {
+    const fn next(&self) -> Option<(u64, u64)> {
         if self.next_section == NO_NEXT_SECTION && self.next_position == NO_NEXT_POSITION {
             None
         } else {
@@ -332,9 +332,9 @@ impl<K: Array> Read for KeyEntry<K> {
     fn read_cfg(buf: &mut impl Buf, _: &Self::Cfg) -> Result<Self, commonware_codec::Error> {
         let key = K::read(buf)?;
         let next_section = u64::read(buf)?;
-        let next_position = u32::read(buf)?;
-        let value_offset = u32::read(buf)?;
-        let value_size = u32::read(buf)?;
+        let next_position = u64::read(buf)?;
+        let value_offset = u64::read(buf)?;
+        let value_size = u64::read(buf)?;
 
         Ok(Self {
             key,
@@ -348,15 +348,15 @@ impl<K: Array> Read for KeyEntry<K> {
 
 impl<K: Array> FixedSize for KeyEntry<K> {
     // key + next_section + next_position + value_offset + value_size
-    const SIZE: usize = K::SIZE + u64::SIZE + u32::SIZE + u32::SIZE + u32::SIZE;
+    const SIZE: usize = K::SIZE + u64::SIZE + u64::SIZE + u64::SIZE + u64::SIZE;
 }
 
 impl<K: Array> OversizedEntryTrait for KeyEntry<K> {
-    fn value_location(&self) -> (u32, u32) {
+    fn value_location(&self) -> (u64, u64) {
         (self.value_offset, self.value_size)
     }
 
-    fn with_location(mut self, offset: u32, size: u32) -> Self {
+    fn with_location(mut self, offset: u64, size: u64) -> Self {
         self.value_offset = offset;
         self.value_size = size;
         self
@@ -372,9 +372,9 @@ where
         Ok(Self {
             key: K::arbitrary(u)?,
             next_section: u64::arbitrary(u)?,
-            next_position: u32::arbitrary(u)?,
-            value_offset: u32::arbitrary(u)?,
-            value_size: u32::arbitrary(u)?,
+            next_position: u64::arbitrary(u)?,
+            value_offset: u64::arbitrary(u)?,
+            value_size: u64::arbitrary(u)?,
         })
     }
 }
@@ -568,7 +568,7 @@ impl<E: Storage + Metrics + Clock, K: Array, V: Codec> Freezer<E, K, V> {
     }
 
     /// Read the latest valid entry from two table slots.
-    fn read_latest_entry(entry1: &Entry, entry2: &Entry) -> Option<(u64, u32, u8)> {
+    fn read_latest_entry(entry1: &Entry, entry2: &Entry) -> Option<(u64, u64, u8)> {
         match (
             !entry1.is_empty() && entry1.is_valid(),
             !entry2.is_empty() && entry2.is_valid(),
